+++ source/app/service-providers/documents/index.ts
@@ -26,7 +26,7 @@ import { ipcMain, app, dialog, type BrowserWindow, type MessageBoxOptions } from
 import { DocumentTree, type DTLeaf } from './document-tree'
 import PersistentDataContainer from '@common/modules/persistent-data-container'
 import { type TabManager } from '@providers/documents/document-tree/tab-manager'
-import { DP_EVENTS, type OpenDocument, DocumentType, BranchNodeJSON, LeafNodeJSON } from '@dts/common/documents'
+import { DP_EVENTS, type OpenDocument, DocumentType, type BranchNodeJSON, type LeafNodeJSON } from '@dts/common/documents'
 import { v4 as uuid4 } from 'uuid'
 import { type Update } from '@codemirror/collab'
 import { ChangeSet, Text } from '@codemirror/state'